requirements.txt
click==8.1.3
greenlet==2.0.2
markdown-it-py==2.2.0
mdurl==0.1.2
psycopg2-binary==2.9.5
pydantic==1.10.7
Pygments==2.14.0
python-decouple==3.4
rich==13.3.3
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a32
sqlmodel==0.0.8
typer==0.7.0
typing_extensions==4.5.0
